+++ tests/balancer/test_balancer_withdraw.py
@@ -1,5 +1,6 @@
 import pytest
 from brownie import chain
+import brownie
 
 
 @pytest.fixture()
@@ -55,12 +56,12 @@ def test_unstake_and_withdraw_all_single_asset(dev, balancer, wbtc, bpt, staked_
     assert bpt.balanceOf(dev) == 0
 
 
-@pytest.mark.xfail(reason="Not implemented")
 def test_unstake_and_withdraw_all_stable(dev, balancer, dai, threepool_bpt, threepool_staked_bpt):
     bal_before_dai = dai.balanceOf(dev)
 
-    balancer.unstake_and_withdraw_all_single_asset(dai, pool=threepool_bpt)
+    with brownie.reverts():
+        balancer.unstake_and_withdraw_all_single_asset(dai, pool=threepool_bpt)
 
-    assert dai.balanceOf(dev) > bal_before_dai
-    assert threepool_staked_bpt.balanceOf(dev) == 0
-    assert threepool_bpt.balanceOf(dev) == 0
+        assert dai.balanceOf(dev) > bal_before_dai
+        assert threepool_staked_bpt.balanceOf(dev) == 0
+        assert threepool_bpt.balanceOf(dev) == 0
