+++ tests/curve/test_curve_deposit.py
@@ -1,4 +1,5 @@
 import pytest
+import brownie
 
 
 def test_deposit_given_amounts(safe, curve, threepool_lptoken, USDC):
@@ -24,7 +25,8 @@ def test_deposit_given_token(safe, curve, threepool_lptoken, USDC):
     assert USDC.balanceOf(safe) < before_bal_usdc
 
 
-@pytest.mark.xfail
 def test_deposit_no_token_specified(curve, threepool_lptoken, USDC):
     amount = 100_000 * 10**USDC.decimals()
-    curve.deposit(threepool_lptoken, amount)
+
+    with brownie.reverts():
+        curve.deposit(threepool_lptoken, amount)
