+++ tests/rari/conftest.py
@@ -1,0 +1,7 @@
+import pytest
+
+
+@pytest.fixture
+def rari(dev):
+    dev.init_rari()
+    return dev.rari
